requirements.txt
streamlit
pandas
numpy
tensorflow-cpu==2.10.1
